requirements.txt
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
annotated-types==0.7.0
attrs==26.1.0
et_xmlfile==2.0.0
frozenlist==1.8.0
idna==3.11
magic-filter==1.0.12
maxapi==0.9.18
multidict==6.7.1
openpyxl==3.1.5
propcache==0.4.1
puremagic==1.30
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.2
typing-inspection==0.4.2
typing_extensions==4.15.0
yarl==1.23.0
